+++ src/boot/efi/meson.build
@@ -258,6 +258,7 @@ libefi_sources = files(
         'devicetree.c',
         'drivers.c',
         'efi-string.c',
+        'efivars.c',
         'export-vars.c',
         'graphics.c',
         'initrd.c',
